+++ src/PhpWord/Shared/XMLWriter.php
@@ -68,9 +68,8 @@ class XMLWriter
             // Create temporary filename
             $this->tempFile = @tempnam($tempFolder, 'xml');
 
-            // Open storage
+            // Fallback to memory when temporary file cannot be used
             if ($this->xmlWriter->openUri($this->tempFile) === false) {
-                // Fallback to memory...
                 $this->xmlWriter->openMemory();
             }
         }
@@ -105,9 +104,16 @@ class XMLWriter
      *
      * @param mixed $function
      * @param mixed $args
+     * @throws \BadMethodCallException
      */
     public function __call($function, $args)
     {
+        // Catch exception
+        if (method_exists($this->xmlWriter, $function) === false) {
+            throw new \BadMethodCallException("Method '{$function}' does not exists.");
+        }
+
+        // Run method
         try {
             @call_user_func_array(array($this->xmlWriter, $function), $args);
         } catch (\Exception $ex) {