+++ src/PhpWord/Style/Cell.php
@@ -144,7 +144,7 @@ class Cell extends Border
      */
     public function getBgColor()
     {
-        if (!is_null($this->shading)) {
+        if ($this->shading !== null) {
             return $this->shading->getFill();
         } else {
             return null;