+++ src/PhpWord/Style/Table.php
@@ -169,7 +169,7 @@ class Table extends Border
      */
     public function getBgColor()
     {
-        if (!is_null($this->shading)) {
+        if ($this->shading !== null) {
             return $this->shading->getFill();
         }