+++ src/PhpWord/Style/TextBox.php
@@ -179,7 +179,7 @@ class TextBox extends Image
         $hasInnerMargins = false;
         $margins = $this->getInnerMargin();
         for ($i = 0; $i < count($margins); $i++) {
-            if (!is_null($margins[$i])) {
+            if ($margins[$i] !== null) {
                 $hasInnerMargins = true;
             }
         }